distillation_in_rl-b0e582cea003cdf38549a73a3f9c0c2c76e1ad63/thesis-code-experiments_run/requirements/server_requirements.txt
# set CUDA_VISIBLE_DEVICES=1 ?
# python3 -m venv <virtual-environment-name>
# source env/bin/activate
# latest torch with cuda: pip3 install torch torchvision torchaudio
# requirememts for running octo
absl-py==2.1.0
# -e git+ssh://git@gitos.rrze.fau.de/utn-machine-intelligence/lab/agents.git@a45eaebb4373c56d7de260af2617362be29fcd1a#egg=agents&subdirectory=../../agents
alembic==1.13.3
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arm_pytorch_utilities==0.4.3
array_record==0.5.1
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
# baselines @ file:///home/piscenco/thesis2/baselines
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
chex==0.1.5
# cleanrl @ file:///home/piscenco/thesis2/cleanrl
click==8.1.7
cloudpickle==3.0.0
colorlog==6.8.2
comm==0.2.2
commonmark==0.9.1
contextlib2==21.6.0
contourpy==1.3.0
cycler==0.12.1
dacite==1.8.1
databricks-sdk==0.32.3
debugpy==1.8.2
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.8
distrax==0.1.5
# dlimp @ file:///home/piscenco/thesis2/dlimp
dm-tree==0.1.8
docker==7.1.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
etils==1.9.4
exceptiongroup==1.2.1
executing==2.0.1
Farama-Notifications==0.0.4
fast_kinematics==0.2.2
fastjsonschema==2.20.0
filelock==3.15.4
flake8==7.1.1
Flask==3.0.3
flatbuffers==24.3.25
flax==0.7.5
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
glfw==2.7.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
graphene==3.3
graphql-core==3.2.4
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.66.1
gunicorn==23.0.0
gym==0.23.1
gym-notices==0.0.8
gymnasium==0.28.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.11.1
hydra-colorlog==1.2.0
hydra-core==1.3.2
identify==2.6.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.4.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.33
jax-cuda12-pjrt==0.4.31
jax-cuda12-plugin==0.4.31
jax-jumpy==1.0.0
jaxlib==0.4.33
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json-numpy==2.1.0
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==2.15.0
kiwisolver==1.4.7
libclang==18.1.1
lxml==5.3.0
Mako==1.3.5
# mani-skill @ file:///home/piscenco/thesis2/ManiSkill
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
ml-collections==0.1.1
ml_dtypes==0.5.0
mlflow==2.16.2
mlflow-skinny==2.16.2
moviepy==1.0.3
mplib==0.1.1
mpmath==1.3.0
msgpack==1.0.8
mujoco==3.2.3
mypy==1.11.2
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nodeenv==1.9.1
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvcc-cu12==12.6.68
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
oauthlib==3.2.2
# octo @ file:///home/piscenco/thesis2/octo
# -e git+ssh://git@gitos.rrze.fau.de/utn-machine-intelligence/student-projects/master-thesis-margarita-piscenco/thesis-code.git@973058e3097097aab8187471a3a3e314a75b7beb#egg=octoplus
omegaconf==2.3.0
opencv-python==4.10.0.84
OpenEXR==3.2.4
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opt-einsum==3.3.0
optax==0.1.4
orbax-checkpoint==0.4.4
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathtools==0.1.2
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plotly==5.24.0
plumbum==1.8.3
pre-commit==3.0.0
proglog==0.1.10
prometheus_client==0.20.0
promise==2.3
prompt_toolkit==3.0.47
protobuf==4.25.5
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==24.7.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
pygame==2.1.0
Pygments==2.18.0
PyOpenGL==3.1.7
pyparsing==3.1.4
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytorch-seed==0.2.0
pytorch_kinematics==0.7.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.0.3
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==11.2.0
rpds-py==0.19.0
rpyc==6.0.1
rsa==4.9
safetensors==0.4.4
sapien==3.0.0b1
scikit-learn==1.5.2
scipy==1.14.1
Send2Trash==1.8.3
sentry-sdk==2.14.0
setproctitle==1.3.3
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.35
sqlparse==0.5.1
stable-baselines3==2.0.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.15.0
tensorflow-graphics==2021.12.3
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.15.0
tensorflow-probability==0.23.0
tensorflow-text==2.15.0
tensorstore==0.1.45
termcolor==2.4.0
terminado==0.18.1
tf_keras==2.15.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
toppra==0.6.0
torch==2.4.1+cu124
torchaudio==2.4.1+cu124
torchvision==0.19.1+cu124
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
transforms3d==0.4.2
trimesh==4.4.9
triton==3.0.0
typeguard==2.13.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tyro==0.5.18
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.3
virtualenv==20.26.3
wandb==0.13.11
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wrapt==1.14.1
zipp==3.20.1
